+++ src/main/java/act/event/LogoutEvent.java
@@ -1,0 +1,39 @@
+package act.event;
+
+/*-
+ * #%L
+ * ACT Framework
+ * %%
+ * Copyright (C) 2014 - 2019 ActFramework
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import act.app.ActionContext;
+import org.osgl.$;
+
+/**
+ * Raised once {@link ActionContext#logout()} is called
+ */
+public class LogoutEvent extends ActEvent {
+    private String identifier;
+
+    public LogoutEvent(String identifier) {
+        this.identifier = $.requireNotNull(identifier);
+    }
+
+    public String identifier() {
+        return identifier;
+    }
+}